# nodejs/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This documentation provides information about all the available endpoints.\n"
  version: "1.0.0"
  title: "NodeJS API"
  contact:
    email: "maksim.kavalenka@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8090"
basePath: "/api"
...
schemes:
- "http"
paths:
...
  /products:
    x-swagger-router-controller: product
    get:
      tags:
      - "product"
      summary: "Get products"
      operationId: "getProducts"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "x-access-token"
        in: "header"
        description: "Authorization header"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ProductsResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
    post:
      tags:
      - "product"
      summary: "Add product"
      operationId: "addProduct"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Product object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/ProductRequest"
      - name: "x-access-token"
        in: "header"
        description: "Authorization header"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ProductResponse"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/BadRequestResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
  /products/{id}:
    x-swagger-router-controller: product
    get:
      tags:
      - "product"
      summary: "Get product"
      operationId: "getProduct"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of product to return"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "Authorization header"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ProductResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
    delete:
      tags:
      - "product"
      summary: "Delete product"
      operationId: "deleteProduct"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of product to delete"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "Authorization header"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/OKResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/NotFoundResponse"
  /products/{id}/reviews:
    x-swagger-router-controller: product
    get:
      tags:
      - "review"
      summary: "Get product's reviews"
      operationId: "getProductReviews"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of product to return reviews for"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "Authorization header"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/ReviewsResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
...
definitions:
...
  Product:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "1f"
      name:
        type: "string"
        example: "Hat"
      brand:
        type: "string"
        example: "HatBrand"
      company:
        type: "string"
        example: "HatCompany"
      price:
        type: "number"
        example: 189.990000000000009094947017729282379150390625
      isbn:
        type: "string"
        example: "3-736-11294-8"
    example:
      price: 189.990000000000009094947017729282379150390625
      isbn: "3-736-11294-8"
      name: "Hat"
      company: false
      _id: "1f"
      brand: "HatBrand"
  ProductRequest:
    type: "object"
    required:
    - "_id"
    - "brand"
    - "company"
    - "isbn"
    - "name"
    - "price"
    properties:
      _id:
        type: "string"
        example: "1f"
      name:
        type: "string"
        example: "Hat"
      brand:
        type: "string"
        example: "HatBrand"
      company:
        type: "string"
        example: "HatCompany"
      price:
        type: "number"
        example: 189.990000000000009094947017729282379150390625
      isbn:
        type: "string"
        example: "3-736-11294-8"
    example:
      price: 189.990000000000009094947017729282379150390625
      isbn: "3-736-11294-8"
      name: "Hat"
      company: false
      _id: "1f"
      brand: "HatBrand"
  ProductResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        example: 200
      message:
        type: "string"
        example: "OK"
      data:
        $ref: "#/definitions/Product"
    example:
      code: 200
      data:
        price: 189.990000000000009094947017729282379150390625
        isbn: "3-736-11294-8"
        name: "Hat"
        company: "HatCompany"
        _id: "1f"
        brand: "HatBrand"
      message: "OK"
  ProductsResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        example: 200
      message:
        type: "string"
        example: "OK"
      data:
        type: "array"
        items:
          $ref: "#/definitions/Product"
    example:
      code: 200
      data:
      - price: 189.990000000000009094947017729282379150390625
        isbn: "3-736-11294-8"
        name: "Hat"
        company: "HatCompany"
        _id: "1f"
        brand: "HatBrand"
      - price: 189.990000000000009094947017729282379150390625
        isbn: "3-736-11294-8"
        name: "Hat"
        company: "HatCompany"
        _id: "1f"
        brand: "HatBrand"
      message: "OK"
  Review:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "8d"
      productId:
        type: "string"
        example: "1f"
      author:
        type: "string"
        example: "Sam"
      text:
        type: "string"
        example: "Product is gorgeous"
    example:
      productId: "1f"
      author: "Sam"
      _id: "8d"
      text: "Product is gorgeous"
  ReviewsResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        example: 200
      message:
        type: "string"
        example: "OK"
      data:
        type: "array"
        items:
          $ref: "#/definitions/Review"
    example:
      code: 200
      data:
      - productId: "1f"
        author: "Sam"
        _id: "8d"
        text: "Product is gorgeous"
      - productId: "1f"
        author: "Sam"
        _id: "8d"
        text: "Product is gorgeous"
      message: "OK"
...
  OKResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        example: 200
      message:
        type: "string"
        example: "OK"
      data:
        type: "string"
        example: "Success"
    example:
      code: 200
      data: "Success"
      message: "OK"
  BadRequestResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        example: 400
      message:
        type: "string"
        example: "Bad Request"
      error:
        type: "array"
        items:
          $ref: "#/definitions/BadRequestResponse_error"
  ForbiddenResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        example: 403
      message:
        type: "string"
        example: "Forbidden"
      error:
        $ref: "#/definitions/ForbiddenResponse_error"
  NotFoundResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        example: 404
      message:
        type: "string"
        example: "Not Found"
      error:
        type: "string"
        example: "Data are not found"
...
  BadRequestResponse_error:
    properties:
      path:
        type: "string"
        example: "attribute"
      message:
        type: "string"
        example: "attribute is required"
  ForbiddenResponse_error:
    properties:
      message:
        type: "string"
        example: "No token provided / invalid signature / jwt expired / Login or password\
          \ is wrong"